requirements.txt
asgiref==3.2.10
Django==3.0.8
django-rest-framework==0.1.0
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
